Move Doctrine driver to new namespace
